cpufreq: cpu-boost: Consider only task load to decide on sync frequency Currently we take the maximum between the source CPU frequency and the calculated frequency based on the migrating task load to decide on the frequency to sync the destination CPU to. This was done to handle short bursts in workloads of tasks which migrated immediately after causing a ramp up on the source CPU. Since their load history wasn't high enough, the destination CPU synced to a lower frequency which wasn't sufficient for the spike in workload.
But as such cases are rare, taking the higher of source and calculated
frequency can lead to destination CPU unnecessarily spending a
considerable amount of time at higher frequencies which in turn can
hurt power.

With this change we make sure only the migrating task load is used
to calculate the sync frequency for destination CPU when load based
syncing is enabled.

Signed-off-by: Girish Ghongdemath <girishsg@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>